+++ src/App.js
@@ -16,12 +16,15 @@ import Table from './components/StockList';
 import { Authstock } from './components/Stock-list provider';
 import { Wallprovider } from './components/Wallet-provider';
 import Demo from './components/Demo';
+import Dashboard from './components/Dashboard';
+import RequiredAuth from './components/RequiredAuth';
 
 
 
 const App = () => {
   return (
     <AnimatePresence mode='wait'>
+       <Wallprovider>
 
       <div className="w-screen sm:h-full md:h-screen flex flex-col bg-primary">
         <Header />
@@ -30,28 +33,28 @@ const App = () => {
           <Authstock>
 
 
-            <Wallprovider>
+           
               <Routes>
                 <Route path="/*" element={<MainContainer />} />
                 <Route path="/createItem" element={<CreateContainer />} />
                 <Route path="/about-us" element={<AboutUs />} />
-                <Route path="/dash" element={<DoughnutChart />} />
+                <Route path="/dash" element={<RequiredAuth><Dashboard /></RequiredAuth>} />
                 <Route path="/news" element={<StocksNews />} />
                 <Route path="/login" element={<Login />} />
                 <Route path="/sign-up" element={<Signup />} />
                 <Route path="/stock-list" element={<Table />} />
                 <Route path="/chart" element={<Chart />} />
-                <Route path="/wallet" element={<WalletContainer />} />
+                <Route path="/wallet" element={<RequiredAuth><WalletContainer /></RequiredAuth>} />
                 <Route path="/demo" element={<Demo />} />
 
               </Routes>
-            </Wallprovider>
           </Authstock>
 
 
         </main>
         <Footer />
       </div>
+      </Wallprovider>
 
     </AnimatePresence>
   );
