+++ src/components/Demo.jsx
@@ -1,200 +1,177 @@
-import { getAuth, GoogleAuthProvider, signInWithPopup } from "firebase/auth";
-import { motion } from "framer-motion";
-// import React, { useState } from 'react';
-import { FiMenu } from 'react-icons/fi';
-import firebase from "firebase/compat/app";
-import { TbLogout } from 'react-icons/tb';
-import { Link } from 'react-router-dom';
-import { actionType } from '../context/reducer';
-import { useStateValue } from '../context/StateProvider';
-import { app } from '../firebase.config';
-import Avatar from '../img/avatar.png';
-import Logo from '../img/logo.png';
-import { FaSearch } from "react-icons/fa";
-import { TfiDashboard } from 'react-icons/tfi'
-import WalletContainer from "./WalletContainer";
-import React, { useEffect, useState } from "react";
-import "firebase/compat/auth";
-import "firebase/compat/firestore";
+import React, { useState, useEffect } from 'react';
+import { useContext } from 'react';
+import axios from 'axios';
+import { Line } from 'react-chartjs-2';
+import Chart from 'chart.js/auto';
+import 'chartjs-adapter-date-fns';
+import { Box, Text, Button } from '@chakra-ui/react';
+import Loader from './Loader'
+const MainContainer = () => {
 
+  
+  let stockname='AAPL'
 
-const firebaseConfig = {
-  // Your Firebase project's configuration object
-  // ...
-
-  apiKey: "AIzaSyBlc8dyS0gFupsNBibnVQYYWd2pIvShbYc",
-  authDomain: "bullishbrains-86d8d.firebaseapp.com",
-  databaseURL: "https://bullishbrains-86d8d-default-rtdb.firebaseio.com",
-  projectId: "bullishbrains-86d8d",
-  storageBucket: "bullishbrains-86d8d.appspot.com",
-  messagingSenderId: "1016988951904",
-  appId: "1:1016988951904:web:d8e6632d2361ef4bee2a05"
-};
-
-function Signup() {
-  const firebaseAuth = getAuth(app);
-  const provider = new GoogleAuthProvider();
-  const [firstName, setFirstName] = useState("");
-  const [lastName, setLastName] = useState("");
-  const [username, setUsername] = useState("");
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [{ user }, dispatch] = useStateValue();
-
+  
+  const [data, setData] = useState(null);
 
   useEffect(() => {
-    if (!firebase.apps.length) {
-      firebase.initializeApp(firebaseConfig);
-    }
-  }, [firebaseConfig]);
+    const fetchData = async () => {
+      const apikey='YYGFPGKNWVMLOYMK'
+      
+      let apiurl=`https://www.alphavantage.co/query?function=TIME_SERIES_DAILY_ADJUSTED&symbol=${stockname}&outputsize=full&apikey=${apikey}`
+      const result = await axios.get(
+        apiurl
+      );
+      setData(result.data['Time Series (Daily)']);
+    };
+    fetchData();
+  }, []);
 
-  const login = async () => {
-    if (!user) {
-      const {
-        user: { providerData },
-      } = await signInWithPopup(firebaseAuth, provider);
-      dispatch({
-        type: actionType.SET_USER,
-        user: providerData[0],
-      });
-
-      if (true) {
-
-        // const user = userCredential.user;
-        await firebase.firestore().collection("users").add({
-          'firstname': providerData['0']['displayName'],
-          'lastname': "",
-          'username': "",
-          'email': providerData['0']['email'],
-          uid: providerData['0']['uid'],
-          wallet: 1000000
-        })
-          .then(() => {
-            console.log("User data stored in Firestore.");
-          })
-          .catch((error) => {
-            console.error("Error creating user: ", error);
-          });
-      }
-
-      localStorage.setItem("user", JSON.stringify(providerData[0]));
-      console.log(providerData);
-    }
+  const chartData = {
+    labels: data ? Object.keys(data).sort() : [],
+    datasets: [
+      {
+        label: 'Reliance Stock Price',
+        data: data
+          ? Object.keys(data)
+            .sort()
+            .map((date) => {
+              return {
+                x: date,
+                y: parseFloat(data[date]['1. open']),
+              };
+            })
+          : [],
+        fill: false,
+        backgroundColor: 'rgba(255, 221, 65, 0.2)',
+        borderColor: '#ffc107',
+        pointRadius: 0,
+      },
+    ],
   };
 
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    firebase.auth().createUserWithEmailAndPassword(email, password)
-  .then((userCredential) => {
-    const user = userCredential.user;
-    // Store user data in Firestore
-    return firebase.firestore().collection("users").add({
-      firstName,
-      lastName,
-      username,
-      email: user.email,
-      uid: user.uid,
-      wallet: 1000000
-    });
-  })
-  .then(() => {
-    console.log("User data stored in Firestore.");
-    // Log in the user
-    return firebase.auth().signInWithEmailAndPassword(email, password);
-  })
-  .then(() => {
-    console.log("User logged in successfully.");
-  })
-  .catch((error) => {
-    console.error("Error creating user: ", error);
-  });
+  const options = {
+    maintainAspectRatio: true,
+    responsive: true,
+    plugins: {
+      legend: {
+        display: true,
+        labels: {
+          font: {
+            size: 16,
+          },
+          color: '#333',
+        },
+      },
+      title: {
+        display: true,
+        text: 'Reliance Industries Ltd. (RELIANCE.BSE)',
+        font: {
+          size: 24,
+          weight: 'bold',
+        },
+        color: '#333',
+        padding: {
+          top: 30,
+          bottom: 10,
+        },
+      },
+      tooltip: {
+        displayColors: false,
+        titleFont: {
+          size: 16,
+          weight: 'bold',
+        },
+        bodyFont: {
+          size: 14,
+        },
+        padding: {
+          top: 10,
+          bottom: 10,
+          left: 10,
+          right: 10,
+        },
+      },
+    },
+    scales: {
+      x: {
+        type: 'time',
+        adapters: {
+          date: require('chartjs-adapter-date-fns'),
+        },
+        time: {
+          unit: 'day',
+          displayFormats: {
+            day: 'MMM d',
+          },
+        },
+        ticks: {
+          color: '#333',
+        },
+        grid: {
+          color: '#eee',
+        },
+      },
+      y: {
+        ticks: {
+          callback: function (value) {
+            return value;
+          },
+          color: '#333',
+        },
+        grid: {
+          color: '#eee',
+        },
+      },
+    },
   };
+  
+
+
 
   return (
-    <div >
-      
-
-      <form onSubmit={handleSubmit} >
-        <div className="mb-4">
-          <label htmlFor="firstName" >
-            First Name:
-          </label>
-          <input
-            type="text"
-            id="firstName"
-            value={firstName}
-            onChange={(e) => setFirstName(e.target.value)}
-            required
-          />
-        </div>
-        <div className="mb-4">
-          <label htmlFor="lastName" >
-            Last Name:
-          </label>
-          <input
-            type="text"
-            id="lastName"
-            value={lastName}
-            onChange={(e) => setLastName(e.target.value)}
-            required
-          />
-        </div>
-        <label htmlFor="username" >
-          Username:
-        </label>
-        <input
-          type="text"
-          id="username"
-          value={username}
-          onChange={(e) => setUsername(e.target.value)}
-          required
-        />
-        <label htmlFor="email">
-          Email:
-        </label>
-        <input
-          type="email"
-          id="email"
-          value={email}
-          onChange={(e) => setEmail(e.target.value)}
-          required
-        />
-        <label htmlFor="password" >
-          Password:
-        </label>
-        <input
-          type="password"
-          id="password"
-          value={password}
-          onChange={(e) => setPassword(e.target.value)}
-          required
-        />
-
-        <p className="text-gray-600 mb-11">
-          Already have an account?{' '}
-          <a
-            href="http://localhost:3000/login"
-          >
-            Sign in
-          </a>{' '}
-        </p>
-          <button
-            type="submit"
-          >
-            Sign Up
-          </button>
-          <button
-            onClick={login}
-            type="submit"
-          >
-            Sign Up with Google
-          </button>
-
-         
-      </form>
+    <div className="grid grid-cols-12 h-screen">
+      <div className="col-span-3 flex items-center justify-center">
+        <Box className="flex flex-col items-center justify-center" borderRadius="md" boxShadow="md" p={4} w="90%">
+          
+          <Box className="flex-col justify-between w-full">
+            <Box className="ml-11 w-1/2 p-4 bg-green-500 rounded-md shadow-md">
+              <Text fontSize="lg" fontWeight="bold" color="#333" mb={2}>
+                Buy
+              </Text>
+              <Button colorScheme="red" size="sm" fontSize="lg" fontWeight="bold" color="#fff" mb={2}>
+                Buy Now
+              </Button>
+            </Box>
+            <Box className="ml-11 w-1/2 p-4 bg-red-500 rounded-md shadow-md mt-5">
+              <Text fontSize="lg" fontWeight="bold" color="#333" mb={2}>
+                Sell
+              </Text>
+              <Button  colorScheme="red" size="sm" fontSize="lg" fontWeight="bold" color="#fff">
+                Sell Now
+              </Button>
+            </Box>
+          </Box>
+        </Box>
+      </div>
+      <div className="col-span-8 flex items-center justify-center">
+        {!data ? (
+          <div className="flex items-center justify-center h-full">
+            <Loader />
+          </div>
+        ) : (
+          <Box className="flex flex-col items-center justify-center mr10" borderRadius="md" boxShadow="md" p={4} w="100%">
+            
+           
+            <Line data={chartData} options={options} className="w-full" />
+          </Box>
+        )}
+      </div>
     </div>
-  );
+  )
+  
+  
+  
 }
 
-export default Signup;
+export default MainContainer
