+++ src/components/Header.jsx
@@ -41,6 +41,12 @@ const Header = () => {
     }
   }, [firebaseConfig]);
 
+  const menuFunct = () => {
+    if(user){
+      setIsMenu(!isMenu);
+    }
+  }
+
   // const login = async () => {
   //   if (!user) {
   //     const {
@@ -178,6 +184,7 @@ const Header = () => {
                 src={user ? user.photoURL : Avatar}
                 className="w-10 min-w-[40px] h-10 min-h-[40px] drop-shadow-xl cursor-pointer rounded-full"
                 alt="userprofile"
+                onClick={menuFunct}
               />
             </Link>
